# swagger-spec.json
{"openapi":"3.0.0","paths":{"/teams":{"get":{"operationId":"TeamController_getAllTeams","parameters":[{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"liga","required":false,"in":"query","schema":{"type":"string"}},{"name":"okreg","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Pobieranie drużyn wraz z opcjonalnym filtorwaniem po nazwie","tags":["teams"]}},"/teams/search":{"get":{"operationId":"TeamController_searchTeams","parameters":[{"name":"name","required":true,"in":"query","description":"Fragment nazwy drużyny","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Szukaj drużyn po nazwie","tags":["teams"]}},"/teams/filter":{"get":{"operationId":"TeamController_filterteam","parameters":[{"name":"liga","required":true,"in":"query","schema":{"type":"string"}},{"name":"okreg","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Filtruj drużyny po lidze i okręgu","tags":["teams"]}},"/teams/filter-extended":{"get":{"operationId":"TeamController_filterTeamWithGroup","parameters":[{"name":"liga","required":true,"in":"query","schema":{"type":"string"}},{"name":"okreg","required":true,"in":"query","schema":{"type":"string"}},{"name":"grupa","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Filtruj drużyny po lidze, okręgu i grupie","tags":["teams"]}},"/teams/{id}":{"get":{"operationId":"TeamController_getTeamById","parameters":[{"name":"id","required":true,"in":"path","description":"Unikalne Id danej drużyny ","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Pobieranie drużyny po jej id","tags":["teams"]}},"/matches/group/{group_id}":{"get":{"operationId":"MatchController_findByGroup","parameters":[{"name":"group_id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":" Pobierze wszsytkie mecze dla konkretnej grupy","tags":["matches"]}},"/matches/team/{teamId}":{"get":{"operationId":"MatchController_findByTeam","parameters":[{"name":"teamId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Pobierz wszystkie mecze dla konkretnej drużyny","tags":["matches"]}},"/matches/{id}":{"get":{"operationId":"MatchController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Pobierz mecz po ID","tags":["matches"]}},"/matches":{"get":{"operationId":"MatchController_findAll","parameters":[],"responses":{"200":{"description":""}},"summary":"Pobierz wszystkie mecze","tags":["matches"]}},"/standings/{group_id}/calculate":{"get":{"operationId":"StandingsController_calculate","parameters":[{"name":"group_id","required":true,"in":"path","description":"ID grupy ligowej","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Przelicz tabelę ligową dla danej grupy","tags":["Standings"]}},"/standings/{group_id}":{"get":{"operationId":"StandingsController_get","parameters":[{"name":"group_id","required":true,"in":"path","description":"ID grupy ligowej","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Pobierz aktualną tabelę ligową dla grupy","tags":["Standings"]}},"/group":{"post":{"operationId":"GroupController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGroupDto"}}}},"responses":{"201":{"description":""}},"tags":["Group"]},"get":{"operationId":"GroupController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Group"]}},"/group/{id}":{"get":{"operationId":"GroupController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Group"]},"patch":{"operationId":"GroupController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGroupDto"}}}},"responses":{"200":{"description":""}},"tags":["Group"]},"delete":{"operationId":"GroupController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Group"]}},"/season":{"post":{"operationId":"SeasonController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSeasonDto"}}}},"responses":{"201":{"description":""}},"tags":["Season"]},"get":{"operationId":"SeasonController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Season"]}},"/season/{id}":{"get":{"operationId":"SeasonController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Season"]},"patch":{"operationId":"SeasonController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSeasonDto"}}}},"responses":{"200":{"description":""}},"tags":["Season"]},"delete":{"operationId":"SeasonController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Season"]}}},"info":{"title":"Meczownik API","description":"API do zarządzania drużynami i meczami","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateGroupDto":{"type":"object","properties":{}},"UpdateGroupDto":{"type":"object","properties":{}},"CreateSeasonDto":{"type":"object","properties":{}},"UpdateSeasonDto":{"type":"object","properties":{}}}}}